Connection Management for Local Proxy
Very simple connection count and denial in the local proxy. If there are more than 50 active requests, the RCLocalProxy waits repeatedly 100 ms, until there is a free slot
before it listens on the port again. Needed to ad a method for in- and decrement the value thread safe before and after handling the request.

refs issue 11