Included HTML Agility Pack.
This library replaces the old HTML parser. It extracts links and references to embedded objects, the yielded results are the same.
The definitions HtmlParser.LinkTagAttributes and HtmlParser.EmbeddedObjectTagAttributes are still used, as I did not find a simpler way to extract references.

The method TranslateToAbsoluteAddress is still used, which is not optimal.