Very rudimentary installer (not working yet). Configuration moved to standard files.
A new installer can be found under /RuralCafe/publish. It copies some files to a very strange directory (not choosable) in the user home directory.
It starts the program, but necessary files are not copies, so that nothing works yet.

In order to have an easier configuration and to have it also when deploying (installer), the recommended way of Visual Studio has been followed.
The config.txt has been removed and settings are now changed the easiest way in Visual Studio -> RuralCafe Properties -> Settings. Also look at the
wiki Introduction page. An UI for changing settings will eventually follow.