Removed /bin from version control and moved it to LocalProxy/RuralCafePages/ and blacklist.txt du RuralCafe root Directory.
The files are now copied to bin (works for both Debug and Release) when the project is built. When new files are added (e.g. in trotro), each file has to changed to be copied on build! An IWizard could solve this.

This change made the installer work so far. There is missing a cool icon and some manifest information like publisher, etc.
Please feel free to configure that in VS under RuralCafe Properties -> Publish.