Headers for GET Commands passed. Richness now configurable.
The headers are now passed from the local to the remote proxy. Potentially also POST parameters are passed, but this is still untested, as POST does not work yet.

The richness is now configurable for each user. The can use the buttons in TroTro and the setting is saved in the remote proxy. (If there is no setting set, the default richness is used). A RC-specific header for the inter-proxy-protocol needed to be added for this: "RCUserID". The addition and removal of other RC-specific headers will be extremely easy.