Improved redirect handling.
If the remote proxy encounters a redirect, he does not only send the redirect page to the local proxy, but also downloads the actual page and packs it.
The local proxy checks file contents when streaming files from the cache. Like this it can determine, if the page contains "HTTP/1.1 301 ..." and redirect accordingly. The result is fine, you don't even notice the redirect. But the method is bad. As soon as we have metadata included in the cache, peeking into the file won't be necessary any more.