Removed the local proxy's global queue, as it was redundant.
The same information is also in the clientQueueMap. To get the next request, we save the user last served and start looking for unsatisfied request after him. ETA calculation still needs to be adapted to this.

This is useful for serializing the queue (and de-serializing it again), as redundant information is difficult to serialize. When (de-)serialization of the clientQueueMap works, we have the "ReadFromLog"-functionality back.