Streaming through remote proxy and accept GZIP/deflate encoding.
When online (no slow network) we're now streaming through both
proxies. If the files are cacheable, they are also put into the cache.
An improvement would be to parallelize the streaming to cache and client.

We do now always accept GZIP and deflate encoding. The HttpWebResponse
automatically decompresses the content accordingly.