Queued Requests spooling on both client and server.
Still not great with 3 different limits on no threads on local/remote end and the max requests.

Not thoroughly tested, there seems to be some bugs when requesting several things at a time, but works in general.
Also, not sure if these random exceptions are caused by these changes or something else.