Cleaned up some code. The HandleRequest and DispatchRequest methods now both belong to the RequestHandlers respectively. The HandleRequests handle incoming requests. They simply check for correctness of Uris, blacklists, setup constructors, and route to stream/queue. If requests are streamed at this point, they are done. Otherwise, they are queued. Connections are closed in all cases EXCEPT for requests at the remote proxy that are from the local proxy.
The DispatchRequest pulls requests from the queue and dispatches them. At the local proxy, this means requests to the remote proxy 
then unpack. At the remote proxy this means requests to the Internet, afterwards, pack and send. Connections are closed afterwards
by the sender (remote proxy).

Switched everything to using threadpools. The limit on number of threads should be changed in the next patch to be independent
of the number of outstanding requests made by the local proxy and by the remote proxy. I've depricated this temporarily in this patch
because there was a lot going on.